requirements.txt
# Requirements for Moodle Docker Infrastructure Installer
# Python 3.8+

# Core dependencies
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0

# System utilities
psutil>=5.9.0
distro>=1.8.0

# Color output (optional)
colorama>=0.4.6
